benchmarker: improve git checkouts.
The double git-clone behavior (spec[git] -> gitdir -> benchmark) means
that some references can not be easilly tested, as a ref in spec[git]
location will be origin/$refname in the gitdir, and then not present at
all the benchmark checkout.

Replace the benchmark checkout with a git-archive export, and update the
commit identification functions to query with the ref against gitdir.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>